help system and flashbang grenades. also, began using SFINAE for room/object combinations.. might be worth making that a common interface for such a tuple

+++ src/scripts/post-build.sh
@@ -1,4 +1,17 @@
 #!/bin/bash
+function log(){
+	echo "$(date -I): '$1'" >> /tmp/cm-post-build-general
+}
+function clear_log(){
+	echo '' > /tmp/cm-post-build-general
+}
+clear_log
+if [[ -f $PWD/../dont-kill-gc ]]; then
+	log 'Not clearing due to dkgc'
+	echo -e '\e[32m[post-build.sh] not killing gc due to ' $PWD/../dont-kill-gc '\e[0m'
+	exit 0
+fi
+
 GCOUNT=$(pgrep gdb | wc -l)
 if [[ $GCOUNT -gt 0 ]]; then
 	killall gdb
